modified:   src/backend/access/nbtree/nbtpage.c 	modified:   src/backend/access/transam/xlogutils.c 	modified:   src/backend/storage/file/fd.c 	modified:   src/backend/storage/smgr/md.c 	modified:   src/backend/storage/smgr/smgr.c A relative stable release It seems the zero page problem has been sovled
1.adjust the mdread. Some pages might not be written when evicted from memory
(not modified). Hence, they are not in the block_header cache. In old version,
I fake a new block, which is unitialized. However, this is incorrect.
In new version, I distinguish them from the mdextend new block.
Only new block extended by mdextend is faked as new block. Other blocks not in
the cache is network_sync from the standby.
2.adjust the FileSeek problem.

Note
This version can run 10 warehouses on NFS.
The database pages mount of NFS should be configured as async.

Remain:
The index structure between master and slave still has inconsistency.
After the master sync a page from slave, it might find it currpted